Pushed 404 page's html, css, js, and a file structure
I added the 404 page's html, css, and js for the project. I also created a simple file structure for the project.

NOTE: I've based the page's js and css references off this file structure so if there's going to be any changes to it, let me know so I can make the appropriate changes to the 404 page's references.

NOTE 2: If you guys want me to change the page's style, js, html, etc. so it better suits the rest of our project, let me know and I'll get to work on it.